Reverse Time :hourglass_flowing_sand:

+++ Algorithms/Dynamic-Programming-Algorithms/staircase_problem.cpp
@@ -7,16 +7,12 @@
 #include <iostream>
 #include <vector>
 
-auto MatrixMultiplication(const std::vector<std::vector<unsigned>> &A, const std::vector<std::vector<unsigned>> &B)
-{
-    size_t n = A.size(), m = B[0].size(), l = B.size();
+auto MatrixMultiplication(const std::vector<std::vector<unsigned>>& A, const std::vector<std::vector<unsigned>>& B) {
+    unsigned n = A.size(), m = B[0].size(), l = B.size();
     std::vector<std::vector<unsigned>> C(n, std::vector<unsigned>(n));
-    for (size_t i = 0; i < n; ++i)
-    {
-        for (size_t j = 0; j < m; ++j)
-        {
-            for (size_t k = 0; k < l; ++k)
-            {
+    for (unsigned i = 0; i < n; ++i) {
+        for (unsigned j = 0; j < m; ++j) {
+            for (unsigned k = 0; k < l; ++k) {
                 C[i][j] += A[i][k] * B[k][j];
             }
         }
@@ -24,16 +20,13 @@ auto MatrixMultiplication(const std::vector<std::vector<unsigned>> &A, const std
     return C;
 }
 
-auto MatrixExponentiation(std::vector<std::vector<unsigned>> &A, unsigned power)
-{
-    size_t n = A.size();
+auto MatrixExponentiation(std::vector<std::vector<unsigned>>& A, unsigned power) {
+    unsigned n = A.size();
     std::vector<std::vector<unsigned>> B(n, std::vector<unsigned>(n));
-    for (size_t i = 0; i < n; ++i)
-    {
+    for (unsigned i = 0; i < n; ++i) {
         B[i][i] = 1;
     }
-    while (power)
-    {
+    while (power) {
         if (power % 2 == 1)
             B = MatrixMultiplication(B, A);
         A = MatrixMultiplication(A, A);
@@ -42,16 +35,13 @@ auto MatrixExponentiation(std::vector<std::vector<unsigned>> &A, unsigned power)
     return B;
 }
 
-unsigned StaircaseProblem(const unsigned &n, const unsigned &k = 3)
-{
-    std::vector<unsigned> baseWays = {0, 1, 2, 4};
+unsigned StaircaseProblem(const unsigned& n, const unsigned& k = 3) {
+    std::vector<unsigned> baseWays = { 0, 1, 2, 4 };
     if (n < 4)
         return baseWays[n];
     std::vector<std::vector<unsigned>> recurrence(k + 1, std::vector<unsigned>(k + 1));
-    for (unsigned i = 1; i < k; ++i)
-    {
-        for (unsigned j = 1; j <= k; ++j)
-        {
+    for (unsigned i = 1; i < k; ++i) {
+        for (unsigned j = 1; j <= k; ++j) {
             if (i == j - 1)
                 recurrence[i][j] = 1;
         }
@@ -59,15 +49,13 @@ unsigned StaircaseProblem(const unsigned &n, const unsigned &k = 3)
     recurrence[k].assign(k + 1, 1);
     recurrence = MatrixExponentiation(recurrence, n - 1);
     unsigned numWays = 0;
-    for (unsigned j = 1; j <= k; ++j)
-    {
+    for (unsigned j = 1; j <= k; ++j) {
         numWays += recurrence[1][j] * baseWays[j];
     }
     return numWays;
 }
 
-int main()
-{
+int main() {
     unsigned numSteps;
     std::cin >> numSteps;
     std::cout << StaircaseProblem(numSteps) << std::endl;
