+++ Algorithms/Mathematical-Algorithms/Linear-Algebra/matrix_exponentiation.cpp
@@ -8,16 +8,12 @@
 #include <vector>
 
 template <typename T>
-auto MatrixProduct(const std::vector<std::vector<T>> &A, const std::vector<std::vector<T>> &B)
-{
-    size_t n = A.size(), m = B[0].size(), l = B.size();
+auto MatrixProduct(const std::vector<std::vector<T>>& A, const std::vector<std::vector<T>>& B) {
+    unsigned n = A.size(), m = B[0].size(), l = B.size();
     std::vector<std::vector<T>> C(n, std::vector<T>(n));
-    for (size_t i = 0; i < n; ++i)
-    {
-        for (size_t j = 0; j < m; ++j)
-        {
-            for (size_t k = 0; k < l; ++k)
-            {
+    for (unsigned i = 0; i < n; ++i) {
+        for (unsigned j = 0; j < m; ++j) {
+            for (unsigned k = 0; k < l; ++k) {
                 C[i][j] += A[i][k] * B[k][j];
             }
         }
@@ -26,11 +22,9 @@ auto MatrixProduct(const std::vector<std::vector<T>> &A, const std::vector<std::
 }
 
 template <typename T>
-void MatrixExponentiation(std::vector<std::vector<T>> A, std::vector<std::vector<T>> &B, unsigned power)
-{
-    size_t n = A.size();
-    while (power)
-    {
+void MatrixExponentiation(std::vector<std::vector<T>> A, std::vector<std::vector<T>>& B, unsigned power) {
+    unsigned n = A.size();
+    while (power) {
         if (power % 2 == 1)
             B = MatrixProduct(B, A);
         A = MatrixProduct(A, A);
@@ -38,27 +32,21 @@ void MatrixExponentiation(std::vector<std::vector<T>> A, std::vector<std::vector
     }
 }
 
-int main()
-{
+int main() {
     unsigned n, power;
     std::cin >> n >> power;
     std::vector<std::vector<double>> matrix(n, std::vector<double>(n)), exponent(n, std::vector<double>(n));
-    for (std::vector<double> &row : matrix)
-    {
-        for (double &element : row)
-        {
+    for (std::vector<double>& row : matrix) {
+        for (double& element : row) {
             std::cin >> element;
         }
     }
-    for (size_t i = 0; i < n; ++i)
-    {
+    for (unsigned i = 0; i < n; ++i) {
         exponent[i][i] = 1;
     }
     MatrixExponentiation(matrix, exponent, power);
-    for (std::vector<double> &row : exponent)
-    {
-        for (double &element : row)
-        {
+    for (std::vector<double>& row : exponent) {
+        for (double& element : row) {
             std::cout << element << ' ';
         }
         std::cout << std::endl;
