+++ Algorithms/Search-Algorithms/matrix_search.cpp
@@ -8,34 +8,29 @@
 #include <vector>
 
 template <typename T>
-auto MatrixSearch(const std::vector<std::vector<T>> &matrix, const T &key)
-{
-    size_t n = matrix.size(), m = matrix[0].size();
+auto MatrixSearch(const std::vector<std::vector<T>>& matrix, const T& key) {
+    unsigned n = matrix.size(), m = matrix[0].size();
     if (key < matrix[0][0] || key > matrix[n - 1][m - 1])
         return std::make_pair(-1, -1);
-    size_t row = 0, col = m - 1;
-    while (row <= n - 1 && col >= 0)
-    {
+    unsigned row = 0, col = m - 1;
+    while (row <= n - 1 && col >= 0) {
         if (matrix[row][col] < key)
             ++row;
         else if (matrix[row][col] > key)
             --col;
         else
-            return {row, col};
+            return { row, col };
     }
     return std::make_pair(-1, -1);
 }
 
-int main()
-{
+int main() {
     int key;
     unsigned n, m;
     std::cin >> n >> m >> key;
     std::vector<std::vector<int>> matrix(n, std::vector<int>(m));
-    for (std::vector<int> &row : matrix)
-    {
-        for (int &element : row)
-        {
+    for (std::vector<int>& row : matrix) {
+        for (int& element : row) {
             std::cin >> element;
         }
     }
