+++ Algorithms/Sorting-Algorithms/Comparison-Sorts/bead_sort.cpp
@@ -7,38 +7,30 @@
 #include <iostream>
 #include <vector>
 
-void BeadSort(std::vector<unsigned> &array)
-{
+void BeadSort(std::vector<unsigned>& array) {
     int max = 0;
-    for (const unsigned &element : array)
-    {
+    for (const unsigned& element : array) {
         if (element > max)
             max = element;
     }
-    size_t n = array.size();
+    unsigned n = array.size();
     std::vector<bool> beads(max * n);
-    for (size_t i = 0; i < n; ++i)
-    {
-        for (unsigned element = 0; element < array[i]; ++element)
-        {
+    for (unsigned i = 0; i < n; ++i) {
+        for (unsigned element = 0; element < array[i]; ++element) {
             beads[max * i + element] = true;
         }
     }
-    for (unsigned element = 0; element < max; ++element)
-    {
+    for (unsigned element = 0; element < max; ++element) {
         unsigned sum = 0;
-        for (size_t i = 0; i < n; ++i)
-        {
+        for (unsigned i = 0; i < n; ++i) {
             sum += beads[max * i + element];
             beads[max * i + element] = false;
         }
-        for (size_t i = n - sum; i < n; ++i)
-        {
+        for (unsigned i = n - sum; i < n; ++i) {
             beads[max * i + element] = true;
         }
     }
-    for (size_t i = 0; i < n; ++i)
-    {
+    for (unsigned i = 0; i < n; ++i) {
         unsigned element = 0;
         while (element < max && beads[max * i + element])
             ++element;
@@ -46,15 +38,13 @@ void BeadSort(std::vector<unsigned> &array)
     }
 }
 
-int main()
-{
+int main() {
     unsigned element;
     std::vector<unsigned> array;
     while (std::cin >> element)
         array.emplace_back(element);
     BeadSort(array);
-    for (const int &element : array)
-    {
+    for (const int& element : array) {
         std::cout << element << ' ';
     }
     std::cout << std::endl;
