+++ Algorithms/Sorting-Algorithms/Comparison-Sorts/cocktail_shaker_sort.cpp
@@ -8,24 +8,18 @@
 #include <vector>
 
 template <typename T>
-void CocktailShakerSort(std::vector<T> &array)
-{
-    size_t low = 0, high = array.size() - 1, swapped = low;
-    while (low < high)
-    {
-        for (size_t i = high; i > low; --i)
-        {
-            if (array[i - 1] > array[i])
-            {
+void CocktailShakerSort(std::vector<T>& array) {
+    unsigned low = 0, high = array.size() - 1, swapped = low;
+    while (low < high) {
+        for (unsigned i = high; i > low; --i) {
+            if (array[i - 1] > array[i]) {
                 std::swap(array[i - 1], array[i]);
                 swapped = i;
             }
         }
         low = swapped;
-        for (size_t i = low; i < high; ++i)
-        {
-            if (array[i] > array[i + 1])
-            {
+        for (unsigned i = low; i < high; ++i) {
+            if (array[i] > array[i + 1]) {
                 std::swap(array[i], array[i + 1]);
                 swapped = i;
             }
@@ -34,15 +28,13 @@ void CocktailShakerSort(std::vector<T> &array)
     }
 }
 
-int main()
-{
+int main() {
     int element;
     std::vector<int> array;
     while (std::cin >> element)
         array.emplace_back(element);
     CocktailShakerSort(array);
-    for (const int &element : array)
-    {
+    for (const int& element : array) {
         std::cout << element << ' ';
     }
     std::cout << std::endl;
