+++ Algorithms/Sorting-Algorithms/Comparison-Sorts/heapsort.cpp
@@ -8,49 +8,41 @@
 #include <vector>
 
 template <typename T>
-void SiftDown(std::vector<T> &heap, const size_t &i, const size_t &n)
-{
-    size_t i_max = i, left = 2 * i + 1, right = left + 1;
+void SiftDown(std::vector<T>& heap, const unsigned& i, const unsigned& n) {
+    unsigned i_max = i, left = 2 * i + 1, right = left + 1;
     if (left < n && heap[left] > heap[i_max])
         i_max = left;
     if (right < n && heap[right] > heap[i_max])
         i_max = right;
-    if (i_max != i)
-    {
+    if (i_max != i) {
         std::swap(heap[i], heap[i_max]);
         SiftDown(heap, i_max, n);
     }
 }
 
 template <typename T>
-void Heapify(std::vector<T> &array)
-{
-    for (int i = array.size() / 2 - 1; i >= 0; --i)
-    {
+void Heapify(std::vector<T>& array) {
+    for (int i = array.size() / 2 - 1; i >= 0; --i) {
         SiftDown(array, i, array.size());
     }
 }
 
 template <typename T>
-void Heapsort(std::vector<T> &array)
-{
+void Heapsort(std::vector<T>& array) {
     Heapify(array);
-    for (size_t i = array.size() - 1, n = array.size(); i > 0; --i)
-    {
+    for (unsigned i = array.size() - 1, n = array.size(); i > 0; --i) {
         std::swap(array[0], array[i]);
         SiftDown(array, 0, --n);
     }
 }
 
-int main()
-{
+int main() {
     int element;
     std::vector<int> array;
     while (std::cin >> element)
         array.emplace_back(element);
     Heapsort(array);
-    for (const int &element : array)
-    {
+    for (const int& element : array) {
         std::cout << element << ' ';
     }
     std::cout << std::endl;
