+++ Algorithms/Sorting-Algorithms/Embarrassingly-Inefficient-Sorts/slowsort.cpp
@@ -8,11 +8,10 @@
 #include <vector>
 
 template <typename T>
-void Slowsort(std::vector<T> &array, const size_t &low, const size_t &high)
-{
+void Slowsort(std::vector<T>& array, const unsigned& low, const unsigned& high) {
     if (low >= high)
         return;
-    size_t mid = low + (high - low) / 2;
+    unsigned mid = low + (high - low) / 2;
     Slowsort(array, low, mid);
     Slowsort(array, mid + 1, high);
     if (array[mid] > array[high])
@@ -20,15 +19,13 @@ void Slowsort(std::vector<T> &array, const size_t &low, const size_t &high)
     Slowsort(array, low, high - 1);
 }
 
-int main()
-{
+int main() {
     int element;
     std::vector<int> array;
     while (std::cin >> element)
         array.emplace_back(element);
     Slowsort(array, 0, array.size() - 1);
-    for (const int &element : array)
-    {
+    for (const int& element : array) {
         std::cout << element << ' ';
     }
     std::cout << std::endl;
